+++ tests/IntegrationTests/DI/Fixtures/BeanInjectionTest/NotFoundVarAnnotation.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * PHP-DI
+ *
+ * @link      http://mnapoli.github.com/PHP-DI/
+ * @copyright 2012 Matthieu Napoli (http://mnapoli.fr/)
+ * @license   http://www.opensource.org/licenses/mit-license.php MIT (see the LICENSE file)
+ */
+
+namespace IntegrationTests\DI\Fixtures\BeanInjectionTest;
+
+use \DI\Annotations\Inject;
+
+/**
+ * Fixture class
+ */
+class NotFoundVarAnnotation {
+
+	/**
+	 * @Inject
+	 * @var this_is_a_non_existent_class
+	 */
+	public $class2;
+
+}